# backend/api-spec.yaml
openapi: 3.0.0
info:
  title: ArcMarket API
  description: API specification for ArcMarket NFT Marketplace
  version: 1.0.0
  contact:
    name: ArcMarket Team
    url: https://arcmarket.io
servers:
  - url: https://api.arcmarket.io/v1
    description: Production server
  - url: http://localhost:3001/v1
    description: Development server
components:
  schemas:
    NFT:
      type: object
      required:
        - id
        - tokenId
        - contractAddress
        - name
        - image
        - collection
        - owner
      properties:
        id:
          type: string
          description: Internal NFT ID
        tokenId:
          type: string
          description: On-chain token ID
        contractAddress:
          type: string
          description: NFT contract address
        name:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        collection:
          $ref: '#/components/schemas/Collection'
        owner:
          type: string
          description: Owner wallet address
        creator:
          type: string
          description: Creator wallet address
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
        price:
          type: string
          description: Current listing price in USDC
        listingId:
          type: string
          description: Active listing ID if listed
    Collection:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        banner:
          type: string
          format: uri
        verified:
          type: boolean
        creator:
          type: string
        creatorAddress:
          type: string
        stats:
          $ref: '#/components/schemas/CollectionStats'
        socials:
          $ref: '#/components/schemas/Socials'
    CollectionStats:
      type: object
      properties:
        floorPrice:
          type: string
        totalVolume:
          type: string
        owners:
          type: integer
        items:
          type: integer
        listed:
          type: integer
        royalty:
          type: number
        change24h:
          type: object
          properties:
            floorPrice:
              type: number
            volume:
              type: number
            sales:
              type: number
    Attribute:
      type: object
      required:
        - trait_type
        - value
      properties:
        trait_type:
          type: string
        value:
          type: string
        rarity:
          type: number
          description: Percentage rarity (0-100)
...
    Socials:
      type: object
      properties:
        twitter:
          type: string
        discord:
          type: string
        website:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet-Signature
paths:
...
  /collection/{slug}:
    get:
      summary: Get collection details
      tags:
        - Collection
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Collection object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        404:
          description: Collection not found
  /collection/{slug}/nfts:
    get:
      summary: Get NFTs in collection
      tags:
        - Collection
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_low, price_high, recently_listed, recently_created, most_viewed, ending_soon]
      responses:
        200:
          description: List of NFTs
          content:
            application/json:
              schema:
                type: object
                properties:
                  nfts:
                    type: array
                    items:
                      $ref: '#/components/schemas/NFT'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
  /collection/{slug}/stats:
    get:
      summary: Get collection statistics
      tags:
        - Collection
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Collection statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionStats'

  # Offer Endpoints
...